Refactor test files (#5)
Refactor test files: 
* Moved test config files into test/config directory
* Split test_examples.py into multiple files. 
* Added new 'actors' test directory.
