+++ test/test_actor_utils.py
@@ -1,3 +1,5 @@
+"""Test actors.utils module"""
+
 import threading
 import time
